send notification almost done, refactored for CA
