+++ src/main/java/use_case/send_notification/SendNotificationDataAccessInterface.java
@@ -1,9 +1,20 @@
 package use_case.send_notification;
 
+import entity.Assignment;
 import entity.Course;
+import entity.User;
+
+import javax.mail.MessagingException;
+import javax.mail.Session;
+import javax.mail.internet.MimeMessage;
 
 public interface SendNotificationDataAccessInterface {
-    void sendNotification();
+
+    Session setupServerProperties();
+
+    MimeMessage draftEmail(Session session, User user, Assignment assignment, Course course) throws MessagingException;
+
+    void sendNotification(Session session, MimeMessage email) throws MessagingException;
 
 }
