+++ src/main/java/use_case/send_notification/SendNotificationInputBoundary.java
@@ -1,5 +1,7 @@
 package use_case.send_notification;
 
+import javax.mail.MessagingException;
+
 /**
  * The Send Notification Use Case.
  */
@@ -8,5 +10,5 @@ public interface SendNotificationInputBoundary {
      * Execute the Send Notification Use Case.
      * @param sendNotificationInputData the input data for this use case
      */
-    void execute(SendNotificationInputData sendNotificationInputData);
+    void execute(SendNotificationInputData sendNotificationInputData) throws MessagingException;
 }
