+++ src/main/java/use_case/send_notification/SendNotificationOutputBoundary.java
@@ -17,5 +17,4 @@ public interface SendNotificationOutputBoundary {
      */
     void prepareFailView(SendNotificationOutputData outputData, String errorMessage);
 
-
 }
